Create a JNI wrapper on the c++ SketchEncrypter, and provide an example on how to use it.
The interface is mostly using byte[] in java, and std::string in c++.
Techniqueluy, we could use the proto and vector<> directly, but the
swig conversion is hard to configure. If this is google internal code,
that would be much easier, since we could take advantage of
base/swig/google.swig.

Change-Id: Ic9a8ebc9371bd9622f870065fce9888a85560a26