Remove wfa directory only in C++ test code.
Right now, the C++ implementation is in: //src/main/cc/...

But the tests are in: //src/test/cc/wfa/...

This CL makes it consistent by moving the tests. At a later time, we can
add /wfa/ back in if we think it should all have that, or we can move
all the C++ code to a separate repo and follow more typical C++
conventions.

Bug: 170483724
Change-Id: I55e3938305797a05326074a8a863db4360758d91