Renames StorageAbstractions.kt to ComputationsDatabase.kt (#7)
* Renames StorageAbstractions.kt to ComputationsDatabase.kt

The naming of and in this file are from a time when we refered to
relational database storage and BLOb sotrage. The terminology changed
to Database means relational and Storage is for BLObs, but the classes
here did not fully adopt that convention.

ComputationsDatabase is the main interface of interest in the file
so the file is named after it.

ComputationsRelationalDb was renamed to ComputationsDatabaseTransactor.
This change is to highlight that the interface performs transactional
read/write operations to the computations database. It also was done
to better distinguish it from the ComputationsDatabase which is
the grouping of an interface for read-only operations and an
interface for read/write transactions.

The ReadOnlyComputationsRelationalDb interface was renamed to
ComputationsDatabaseReader.

The ComputationStorageEditToken was moved to a be a child class of
ComputationsDatabaseTransactor. It is only ever used in that context.
It was also renamed to ComputationEditToken.